Classification/requirements.txt
jieba
pandas
scikit-learn
numpy
matplotlib
wordcloud
python-bidi